OMAP:DSS2: Add manager blanking support
Add manager->blank() to disable all overlays on a manager from
all the dss_cache stages.

blank() takes 2 additional arguments beyond the manager.

wait_for_vsync: to wait for vsync before flush returns.
persist: to disable subsequent applies until the display is
turned off/suspended.

if wait_for_vsync is false, blank() will flush the callbacks
in the shadow and displayed stages of the dss cache.  You may
want to use this if there are no further VSYNCS for the
display.  This is safe because the dss_cache is updated for the
blanked overlays, so it will take effect the next time configure()
is called.

Change-Id: Ic4bbd4ef73246ee89bdc8699b4bc09efffafe51f
Signed-off-by: Lajos Molnar <molnar@ti.com>