Hack PVR omaplfb driver to play nice with DSSCOMP on OMAP3
OMAP3 lacks TILER, which PVR assumes the device has if it has DSSCOMP.
Unlike on OMAP4, DSSCOMP on OMAP3 doesn't interact with the PVR driver
at all, so it should in principle be safe to stub out all of the
DSSCOMP-dependent code in the PVR driver so that we fall back on the
standard code paths.

Credit to fat-tire for the suggestion.